chore: Use standard project layout
