+++ lib/quinn.js
@@ -1,0 +1,104 @@
+/*
+ * Copyright (c) 2019, Groupon, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * Redistributions of source code must retain the above copyright notice,
+ * this list of conditions and the following disclaimer.
+ *
+ * Redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution.
+ *
+ * Neither the name of GROUPON nor the names of its contributors may be
+ * used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS
+ * IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED
+ * TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+ * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+ * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014 Jan Krems <jan.krems@gmail.com>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+'use strict';
+
+const respond = require('./respond');
+
+const NOT_FOUND = new Buffer('Not Found\n', 'utf8');
+const INTERNAL_ERROR = new Buffer('Internal Server Error\n', 'utf8');
+
+function sendNotFound(res) {
+  res.statusCode = 404;
+  res.end(NOT_FOUND);
+}
+
+function sendFatalError(res, err) {
+  try {
+    res.statusCode = 500;
+    res.end(INTERNAL_ERROR);
+  } catch (_) {}
+  return Promise.reject(err);
+}
+
+function runApplication(handler, req, res) {
+  return Promise.resolve(req)
+    .then(handler)
+    .then(function(vres) {
+      if (vres === undefined) return;
+
+      return new Promise(function(resolve, reject) {
+        res.on('error', reject);
+        res.on('finish', function() { resolve(vres); });
+        vres.pipe(res);
+      });
+    });
+}
+
+function createApp(handler) {
+  return function(req, res) {
+    return runApplication(handler, req, res)
+      .then(function(result) {
+        if (result === undefined) return sendNotFound(res);
+        return result;
+      })
+      .then(null, function(err) { return sendFatalError(res, err); });
+  };
+}
+
+module.exports = createApp;
+createApp['default'] = createApp;
+createApp.createApp = createApp;
+createApp.respond = respond;
+createApp.runApplication = runApplication;
